[MOV] product: Re-organized directory structure.
