+++ addons/product/data/product_data.xml
