+++ addons/product/data/product_demo.xml
