+++ addons/product/models/__init__.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-import models
-import report
-import wizard
+import partner
+import pricelist
+import product
+import res_company
+import res_config
