+++ addons/product/models/partner.py
