+++ addons/product/models/res_config.py
