+++ addons/product/views/partner_views.xml
