+++ addons/product/views/pricelist_views.xml
