+++ addons/product/views/product_report.xml
