+++ addons/product/views/product_views.xml
