+++ addons/product/views/res_config_views.xml
